requirements.txt
asgiref==3.3.4
certifi==2021.5.30
chardet==4.0.0
Django==3.2.4
django-ipware==3.0.2
django-rest==0.8.7
djangorestframework==3.12.4
flake8==3.9.2
idna==2.10
isort==5.9.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
types-requests==2.25.0
typing-extensions==3.10.0.0
urllib3==1.26.5
